refactor: add checkAccess util func for fs tasks
