+++ src/fs/copy.js
@@ -1,30 +1,27 @@
-import { access, cp } from 'node:fs';
-import { throwError } from './throwError.js';
+import { cp } from 'node:fs';
+import { checkAccess } from './utils/checkAccess.js';
+import { throwError } from './utils/throwError.js';
 
 const copy = async () => {
   const path = 'src/fs/';
   const dir = path + 'files';
   const newDir = path + 'files_copy';
 
-  access(dir, (err) => {
-    if (err) {
-      throwError();
-    } else {
-      access(newDir, (err) => {
-        if (err) {
-          cp(dir, newDir, { recursive: true }, (error) => {
-            if (error) {
-              console.error(error);
-            } else {
-              console.log('Folder copied successfully');
-            }
-          });
-        } else {
-          throwError();
-        }
-      })
-    }
-  })
+  const callback = () => {
+    cp(dir, newDir, { recursive: true }, (error) => {
+      if (error) {
+        console.error(error);
+      } else {
+        console.log('Folder copied successfully');
+      }
+    });
+  }
+
+  const checkNewDir = () => {
+    checkAccess(newDir, throwError, callback);
+  }
+
+  checkAccess(dir, checkNewDir, throwError);
 };
 
 await copy();
