+++ src/fs/delete.js
@@ -1,22 +1,21 @@
-import { access, unlink } from 'node:fs';
-import { throwError } from './throwError.js';
+import { unlink } from 'node:fs';
+import { checkAccess } from './utils/checkAccess.js';
+import { throwError } from './utils/throwError.js';
 
 const remove = async () => {
   const path = 'src/fs/files/fileToRemove.txt';
 
-  access(path, (err) => {
-    if (err) {
-      throwError();
-    } else {
-      unlink(path, (error) => {
-        if (err) {
-          console.error(error);
-        } else {
-          console.log('File was deleted');
-        }
-      })
-    }
-  })
+  const callback = () => {
+    unlink(path, (error) => {
+      if (error) {
+        console.error(error);
+      } else {
+        console.log('File was deleted');
+      }
+    })
+  };
+
+  checkAccess(path, callback, throwError);
 };
 
 await remove();
