+++ src/fs/list.js
@@ -1,22 +1,21 @@
-import { access, readdir } from 'node:fs';
-import { throwError } from './throwError.js';
+import { readdir } from 'node:fs';
+import { checkAccess } from './utils/checkAccess.js';
+import { throwError } from './utils/throwError.js';
 
 const list = async () => {
   const path = 'src/fs/files/';
 
-  access(path, (err) => {
-    if (err) {
-      throwError();
-    } else {
-      readdir(path, (error, files) => {
-        if (error) {
-          console.error(error);
-        } else {
-          console.log(files.join(',\n'));
-        }
-      })
-    }
-  })
+  const callback = () => {
+    readdir(path, (error, files) => {
+      if (error) {
+        console.error(error);
+      } else {
+        console.log(files.join(',\n'));
+      }
+    })
+  };
+
+  checkAccess(path, callback, throwError);
 };
 
 await list();
