+++ src/fs/read.js
@@ -1,23 +1,22 @@
-import { access, readFile } from 'node:fs';
-import { throwError } from './throwError.js';
+import { readFile } from 'node:fs';
+import { checkAccess } from './utils/checkAccess.js';
+import { throwError } from './utils/throwError.js';
 
 const read = async () => {
   const path = 'src/fs/files/fileToRead.txt';
   const encoding = 'utf8';
 
-  access(path, (err) => {
-    if (err) {
-      throwError();
-    } else {
-      readFile(path, encoding, (err, data) => {
-        if (err) {
-          console.error(err);
-        } else {
-          console.log(data);
-        }
-      })
-    }
-  })
+  const callback = () => {
+    readFile(path, encoding, (err, data) => {
+      if (err) {
+        console.error(err);
+      } else {
+        console.log(data);
+      }
+    })
+  }
+
+  checkAccess(path, callback, throwError);
 };
 
 await read();
