+++ src/fs/rename.js
@@ -1,5 +1,6 @@
-import { access, rename as renameFile } from 'node:fs';
-import { throwError } from './throwError.js';
+import { rename as renameFile } from 'node:fs';
+import { checkAccess } from './utils/checkAccess.js';
+import { throwError } from './utils/throwError.js';
 
 const rename = async () => {
   const path = 'src/fs/files/';
@@ -7,26 +8,22 @@ const rename = async () => {
   const newFile = 'properFilename.md';
   const oldFilePath = path + oldFile;
   const newFilePath = path + newFile;
-  
-  access(oldFilePath, (err) => {
-    if (err) {
-      throwError();
-    } else {
-      access(newFilePath, (err) => {
-        if (err) {
-          renameFile(oldFilePath, newFilePath, (error) => {
-            if (error) {
-              console.log(error);
-            } else {
-              console.log('File was renamed');
-            }
-          })
-        } else {
-          throwError();
-        }
-      })
-    }
-  })
+
+  const callback = () => {
+    renameFile(oldFilePath, newFilePath, (error) => {
+      if (error) {
+        console.log(error);
+      } else {
+        console.log('File was renamed');
+      }
+    })
+  };
+
+  const checkNewFilePath = () => {
+    checkAccess(newFilePath, throwError, callback);
+  }
+
+  checkAccess(oldFilePath, checkNewFilePath, throwError);
 };
 
 await rename();
